refactor: AppState 위치 이동 및 BottomNavTab을 이용한 라우트 계산 로직 개선
- AppState를 navigation 패키지로 이동
- BottomNavTab.entries를 사용하여 currentRoute 계산 로직 리팩토링
- GlobalApplication에 Kakao SDK 초기화 관련 TODO 추가
- AppNavigation의 불필요한 import 제거
