+++ app/src/main/java/com/afternote/afternote_fe/GlobalApplication.kt
@@ -5,6 +5,6 @@ import android.app.Application
 class GlobalApplication : Application() {
     override fun onCreate() {
         super.onCreate()
-//        KakaoSdk.init(this, "{네이티브_APP_KEY}")
+// TODO:KakaoSdk.init(this, "{네이티브_APP_KEY}")
     }
 }
