+++ app/src/main/java/com/afternote/afternote_fe/navigation/AppState.kt
@@ -1,4 +1,4 @@
-package com.afternote.afternote_fe
+package com.afternote.afternote_fe.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Stable
@@ -9,6 +9,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.afternote.core.ui.Route
+import com.afternote.core.ui.component.bottombar.BottomNavTab
 
 // TODO:검토
 @Stable
@@ -23,13 +24,9 @@ class AppState(
     // 바텀바가 선택해야 할 현재 라우트 계산
     val currentRoute: Route?
         @Composable get() =
-            when {
-                currentDestination?.hasRoute(Route.Home::class) == true -> Route.Home
-                currentDestination?.hasRoute(Route.MindRecord::class) == true -> Route.MindRecord
-                currentDestination?.hasRoute(Route.TimeLetter::class) == true -> Route.TimeLetter
-                currentDestination?.hasRoute(Route.Afternote::class) == true -> Route.Afternote
-                else -> null
-            }
+            BottomNavTab.entries
+                .firstOrNull { currentDestination?.hasRoute(it.route::class) == true }
+                ?.route
 
     // 바텀바 아이템 클릭 시 이동하는 네비게이션 로직
     fun navigateToBottomBarRoute(route: Route) {
